chore(1-3342): extract into hook, use new endpoint if flag on
